Restructuring of app and removing loader
