+++ package.json
@@ -21,6 +21,7 @@
     "react-dom": "^17.0.2",
     "react-feather": "^2.0.9",
     "react-leaflet": "^3.2.2",
+    "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
     "react-search-autocomplete": "^5.4.2",
     "react-snowfall": "^1.1.0",
