+++ src/Views/TripPlannerPage/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const TripPlannerPage = () => {
+	console.log('working');
+	return <div>ALL THE components for trip planner will go here</div>;
+};
+
+export default TripPlannerPage;
